# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  models.ChangePasswordInput:
    properties:
      newPassword:
        type: string
      oldPassword:
        type: string
    type: object
...
  models.OrderRecord:
    properties:
      artist:
        type: string
      categoryName:
        type: string
      createdAt:
        type: string
      date:
        type: string
      eventId:
        type: string
      eventTitle:
        type: string
      id:
        type: string
      orderCode:
        type: string
      paymentMethod:
        type: string
      qrCode:
        type: string
      quantity:
        type: integer
      status:
        type: string
      totalPrice:
        type: number
      userEmail:
        type: string
      userId:
        type: string
      userName:
        type: string
      venue:
        type: string
    type: object
...
  models.RefundRequestRecord:
    properties:
      accountHolder:
        type: string
      accountNumber:
        type: string
      adminNote:
        type: string
      bankName:
        type: string
      categoryName:
        type: string
      createdAt:
        type: string
      eventTitle:
        description: Joined details for UI/Admin
        type: string
      id:
        type: string
      orderCode:
        type: string
      orderId:
        type: string
      quantity:
        type: integer
      reason:
        type: string
      refundAmount:
        type: number
      status:
        type: string
      updatedAt:
        type: string
      userEmail:
        type: string
      userName:
        type: string
    type: object
...
  models.UpdateProfileInput:
    properties:
      name:
        type: string
      phone:
        type: string
    type: object
  models.UserDashboardSummary:
    properties:
      activeRefundsCount:
        type: integer
      pastEventsCount:
        type: integer
      totalTicketsBought:
        type: integer
      upcomingEventsCount:
        type: integer
    type: object
  models.UserRecord:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      isVerified:
        type: boolean
      name:
        type: string
      phone:
        type: string
      role:
        type: string
      updatedAt:
        type: string
    type: object
host: localhost:8082
info:
  contact:
    email: support@symphoniatic.id
    name: SymphoniaTic API Support
  description: Interactive OpenAPI/Swagger UI documentation for SymphoniaTic Concert
    Ticketing Platform.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: SymphoniaTic REST API Documentation
  version: "1.0"
paths:
...
  /user/change-password:
    post:
      consumes:
      - application/json
      description: Mengubah kata sandi user dengan mengonfirmasi kata sandi lama.
      parameters:
      - description: Change Password Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.ChangePasswordInput'
      produces:
      - application/json
      responses:
        "200":
          description: Kata sandi akun Anda berhasil diperbarui
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Kata sandi lama salah / minimal 6 karakter
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Mengubah kata sandi akun
      tags:
      - User Profile
  /user/dashboard-summary:
    get:
      consumes:
      - application/json
      description: Mengambil statistik total tiket dibeli, konser mendatang, konser
        selesai, dan refund aktif.
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mengambil ringkasan statistik akun
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.UserDashboardSummary'
              type: object
      security:
      - BearerAuth: []
      summary: Mengambil ringkasan statistik dashboard user
      tags:
      - User Profile
  /user/orders:
    get:
      consumes:
      - application/json
      description: Mengambil semua riwayat tiket/pesanan milik user yang sedang login.
      parameters:
      - description: Filter status pesanan (ISSUED, CHECKED_IN, REFUNDED, dll)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mengambil riwayat pesanan
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.OrderRecord'
                  type: array
              type: object
        "401":
          description: Sesi tidak valid
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Mengambil riwayat pesanan pengguna
      tags:
      - User Profile
  /user/orders/{orderCode}:
    get:
      consumes:
      - application/json
      description: Mengambil detail spesifik pesanan berdasarkan kode pesanan untuk
        user yang login.
      parameters:
      - description: Order Code
        in: path
        name: orderCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mengambil detail pesanan
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.OrderRecord'
              type: object
        "404":
          description: Pesanan tidak ditemukan
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Mengambil detail 1 pesanan pengguna
      tags:
      - User Profile
  /user/profile:
    put:
      consumes:
      - application/json
      description: Mengubah nama lengkap dan nomor telepon pengguna terautentikasi.
      parameters:
      - description: Update Profile Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.UpdateProfileInput'
      produces:
      - application/json
      responses:
        "200":
          description: Profil berhasil diperbarui
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.UserRecord'
              type: object
        "400":
          description: Nama wajib diisi
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Memperbarui profil pengguna
      tags:
      - User Profile
  /user/refunds:
    get:
      consumes:
      - application/json
      description: Mengambil riwayat pengajuan refund tiket beserta status persetujuan
        dari admin.
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mengambil data pengajuan refund
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.RefundRequestRecord'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Mengambil daftar pengajuan refund pengguna
      tags:
      - User Profile
securityDefinitions:
  BearerAuth:
    description: Type "Bearer <your_token>" to authenticate user/admin endpoints.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"